Generalize ExpandableCard
Remove changing shape and padding options
Use icon from Material icons.
